refactor: move `HtmlReporter` to the correct namespace
