+++ src/Stryker.Core/Stryker.Core/Reporters/ReporterFactory.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Stryker.Core.Baseline.Providers;
 using Stryker.Core.Options;
+using Stryker.Core.Reporters.Html;
 using Stryker.Core.Reporters.Json;
 using Stryker.Core.Reporters.Progress;
 
@@ -28,7 +29,7 @@ namespace Stryker.Core.Reporters
                 { Reporter.ClearText, new ClearTextReporter(options) },
                 { Reporter.ClearTextTree, new ClearTextTreeReporter(options) },
                 { Reporter.Json, new JsonReporter(options) },
-                { Reporter.Html, new HtmlReporter.HtmlReporter(options) },
+                { Reporter.Html, new HtmlReporter(options) },
                 { Reporter.Dashboard, new DashboardReporter(options) },
                 { Reporter.Markdown, new MarkdownSummaryReporter(options) },
                 { Reporter.Baseline, new BaselineReporter(options) }
